added isValidId.js
