+++ controllers/contacts.js
@@ -1,9 +1,9 @@
 const { HttpError, ctrlWrapper } = require('../helpers');
 
-const Contact = require('../models/contact');
+const { Contact } = require('../models/contact');
 
 const listContacts = async (req, res) => {
-  const result = await Contact.find({});
+  const result = await Contact.find();
   res.json(result);
 };
