+++ middlewares/index.js
@@ -1,9 +1,9 @@
 const validateBody = require('./validateBody');
 const checkBody = require('./checkBody');
-const checkField = require('./checkField');
+const isValidId = require('./isValidId');
 
 module.exports = {
   validateBody,
   checkBody,
-  checkField,
+  isValidId,
 };
