+++ routes/api/contacts.js
@@ -2,31 +2,33 @@ const express = require('express');
 
 const ctrl = require('../../controllers/contacts');
 
-const { validateBody, checkBody } = require('../../middlewares');
+const { validateBody, isValidId, checkBody } = require('../../middlewares');
 
-const schemas = require('../../schemas/contacts');
+const schemas = require('../../models/contact');
 
 const router = express.Router();
 
 router.get('/', ctrl.listContacts);
 
-router.get('/:id', ctrl.getContactById);
+router.get('/:id', isValidId, ctrl.getContactById);
 
 router.post('/', checkBody, validateBody(schemas.addSchema), ctrl.addContact);
 
 router.put(
   '/:id',
+  isValidId,
   checkBody,
   validateBody(schemas.addSchema),
   ctrl.updateContact
 );
 
-router.delete('/:id', ctrl.removeContact);
+router.delete('/:id', isValidId, ctrl.removeContact);
 
 router.patch(
   '/:id/favorite',
+  isValidId,
   checkBody,
-  validateBody(schemas.addSchema),
+  validateBody(schemas.updateFavoriteSchema),
   ctrl.updateStatusContact
 );
